Add example files for Rigol DS2072A scope.
Both channels were connected to the calibrator output. The channels were set to
1x (default setup), but probes were 10x, so the square signal amplitude in the
files appears as 300 mV, not 3 V.

Captures 1 through 5 are using both channels using various time bases and
attenuator settings. Captures 6 and 7 use only a single channel.

Files were saved after a single trigger capture, so both .csv and .wfm
should have exactly the same data.

Most screenshots don't actually show the waveform, but the basic settings can
still be seen on them (as well as in the included .txt files). Not showing the
waveform in the screenshot .png seems to be a common bug in these scopes.